Skip unavailable videos when downloading from Playlist
Often YouTube Playlists have some Private videos which cannot be accessed by everyone. When that happens, the bot raises an exception regarding the same. Moreover skipping private videos is not a default behavior of youtube-dl. Therefore I see the idea of skipping the unavailable videos while downloading from a playlist.